Several improvements with logbase_mu
Before, the transformation was made in the top-level function. Then, the output had the coefficients always in log10 scale. This was quite unconvenient for model fitting.

Now, the transformation is down in the down-level function, and only for the calculations. I have also added an attribute to the class with the logbase, so I can keep the numeric value of mu in the right scale. This scale is also shown in the print() method